requirements.txt
# =============================================================================
# requirements.txt — All Python Packages Needed
# =============================================================================
# Install with: pip install -r requirements.txt
#
# Package reference:
#   langchain              → Core LangChain framework (chains, prompts, memory)
#   langchain-community    → Community integrations (PyPDFLoader, FAISS)
#   langchain-groq         → Groq API integration for ChatGroq
#   langchain-huggingface  → HuggingFaceEmbeddings (replaces deprecated community import)
#   faiss-cpu              → FAISS vector store (local, CPU build)
#   sentence-transformers  → all-MiniLM-L6-v2 embedding model
#   pypdf                  → PDF reader used by PyPDFLoader
#   streamlit              → Web UI framework
#   python-dotenv          → Reads GROQ_API_KEY from .env
# =============================================================================

langchain>=0.3,<0.4
langchain-community>=0.3,<0.4
langchain-groq>=0.2,<0.3
langchain-huggingface>=0.1,<0.2
faiss-cpu>=1.8,<2.0
sentence-transformers>=3.0,<4.0
pypdf>=4.0,<5.0
streamlit>=1.35,<2.0
python-dotenv>=1.0,<2.0
msvc-runtime
